pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "runpod-runner"
version = "0.1.0"
description = "Runs ComfyScript on Runpod"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["comfy-script", "Pillow", "requests"]

[project.scripts]
runpod_runner = "runpod_runner.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
